backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0
pydantic==2.7.1
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
scipy==1.13.1
Pillow==10.3.0
deepface==0.0.92
tf-keras==2.16.0
openai-whisper==20231117
torch==2.3.0
torchaudio==2.3.0
aiofiles==23.2.1
python-jose==3.3.0
